move pages into scene directory, update imports
